This should fix most Bush issues but has numerical problems
This commit does the following:
*	reenables topological order caching
*	includes null checking for longest path links in ASP
*	includes Link self-checking in path relaxation
*	NaN-checks bush flows and merge splits
*	rearranges shortest path topological search to check incoming links
*	waits to activate links until all have been explored when improving
*	updates BushMerge splits only when the total flow is > 0
*	keeps longest path as owner of all flow when creating BushMerge
*	disables ASP generation when doing shortest/longest path search
*	adds some numerical checking when altering link flows (problematic)

+++ src/edu/utexas/wrap/util/calc/GapCalculator.java
@@ -38,7 +38,7 @@ public class GapCalculator extends Thread {
 				Map<Node, Double> cache = new HashMap<Node, Double>(graph.numNodes());
 				for (Node d : b.getNodes()) {
 					
-					Float demand = b.getDemand(d);
+					Float demand = b.getDemand(d).floatValue();
 					if (demand > 0.0F) try {
 						denominator += b.getCachedL(d,cache).doubleValue() * demand;
 					} catch (UnreachableException e) {
